+++ tests/test_estimation.py
@@ -1724,6 +1724,14 @@ class TestCoxTimeVaryingFitter():
         assert True
 
 
+    def test_likelihood_ratio_test_against_R(self, ctv, heart):
+        ctv.fit(heart, id_col='id', event_col='event')
+        test_stat, deg_of_freedom, p_value = ctv._compute_likelihood_ratio_test()
+        assert abs(test_stat - 15.1) < 0.1
+        assert abs(p_value - 0.00448) < 0.001
+        assert deg_of_freedom == 4
+
+
     def test_print_summary(self, ctv, heart):
 
         import sys
@@ -1744,6 +1752,8 @@ surgery    -0.6372     0.5288    0.3672 -1.7352 0.0827     -1.3570      0.0825  
 transplant -0.0103     0.9898    0.3138 -0.0327 0.9739     -0.6252      0.6047
 ---
 Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
+
+Likelihood ratio test = 15.111 on 4 df, p=0.00448
 """.strip().split()
             for i in [0, 1, 2, 3, -2, -1, -3, -4, -5]:
                 assert output[i] == expected[i]