+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 #### 0.14.4
  - Adds `baseline_cumulative_hazard_` and `baseline_survival_` to `CoxTimeVaryingFitter`. Because of this, new prediction methods are available. - fixed a bug in `add_covariate_to_timeline` when using `cumulative_sum` with multiple columns.
- - Added `Likelihood ratio test` to `CoxPHFitter.print_summary`
+ - Added `Likelihood ratio test` to `CoxPHFitter.print_summary` and `CoxTimeVaryingFitter.print_summary`
  - New checks in `CoxTimeVaryingFitter` that check for immediate deaths and redundant rows. 
  - New `delay` parameter in `add_covariate_to_timeline`